requirements.txt
livekit-agents==0.12.16
livekit-plugins-openai[vertex]==0.10.19
livekit-plugins-deepgram==0.6.19
livekit-plugins-silero==0.7.4
livekit-plugins-aws==0.1.0
livekit-plugins-cartesia==0.4.8
python-dotenv==1.0.0
livekit==0.19.1
twilio==9.5.0

# play.asi==1.0.0  # Commented out until the correct source is identified
